pyproject.toml
[tool.poetry]
name = "smartnotifications"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8"
python-telegram-bot = "^21.6"
requests = "^2.32.3"
openai = "^1.50.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
